requirements.txt
# requirements.txt

aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.5
discord.py==2.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
motor==3.4.0
multidict==6.0.5
pycparser==2.22
pymongo==4.7.0
python-dotenv==1.0.1
yarl==1.9.4
